ui: Be more robust about clearing the line
Some terminals (like urxvt) will allow you to "backspace" (\x08) past the
beginning of the current line, in to the previous line.  This means that
each update to the status line moves farther and farther *up* the screen.

Instead, prefix each line with "\r" to move the cursor to the beginning of
the line, and pad the new line with whitespace at the end to make it at
least as long as the previous line, in order to overwrite it.  This means
that we must keep track of the length of the previous line.